Move test_gsobject tests into more appropriate files (#372b)
